# strict/swagger.json
{
    "openapi": "3.0.3",
    "paths": {
...
        "/voiture/option/{option}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "option",
                        "schema": {
                            "type": "object",
                            "example": "\"[{\\\"$match\\\":{}}]\""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "VoitureController"
                ]
            }
        },
        "/voiture/{id}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "VoitureController"
                ]
            },
            "delete": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "properties": {
                                "_bsontype": {
                                    "type": "object",
                                    "example": "\"ObjectID\""
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "VoitureController"
                ]
            }
        },
        "/voiture": {
            "post": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "VoitureController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "year": {
                                        "type": "string"
                                    },
                                    "numero": {
                                        "type": "string"
                                    },
                                    "clientId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    },
                                    "modelVoitureId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "VoitureController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "year": {
                                        "type": "string"
                                    },
                                    "numero": {
                                        "type": "string"
                                    },
                                    "clientId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    },
                                    "modelVoitureId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}